+++ docs/court_wizard_bg.wasm.d.ts
@@ -2,14 +2,14 @@
 /* eslint-disable */
 export const memory: WebAssembly.Memory;
 export const main: (a: number, b: number) => number;
-export const __wasm_bindgen_func_elem_125195: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_125174: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_128222: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_128212: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_129881: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_129759: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_129880: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_129882: (a: number, b: number, c: number, d: number) => void;
+export const __wasm_bindgen_func_elem_130062: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_129940: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_125376: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_125355: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_130061: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_130063: (a: number, b: number, c: number, d: number) => void;
+export const __wasm_bindgen_func_elem_128403: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_128393: (a: number, b: number) => void;
 export const __wbindgen_export: (a: number, b: number) => number;
 export const __wbindgen_export2: (a: number, b: number, c: number, d: number) => number;
 export const __wbindgen_export3: (a: number) => void;