+++ src/game/units/mod.rs
@@ -7,6 +7,7 @@ use rand::Rng;
 
 pub(crate) mod archer;
 pub(crate) mod behemoth;
+pub(crate) mod boss;
 pub(in crate::game) mod commander;
 pub(crate) mod components;
 pub(crate) mod constants;